+++ tests/texture_write.rs
@@ -11,7 +11,7 @@ fn texture_2d_write() {
     let texture = glium::texture::Texture2d::new(&display, vec![
         vec![(0u8, 1u8, 2u8), (4u8, 8u8, 16u8)],
         vec![(32u8, 64u8, 128u8), (32u8, 16u8, 4u8)],
-    ]);
+    ]).unwrap();
 
     texture.write(glium::Rect { bottom: 1, left: 1, width: 1, height: 1 },
                   vec![vec![(128u8, 64u8, 2u8)]]);